add tokens to uniprice
